fix: contracts 模块 P0-P3 全面优化
P0: 文件清理信号 + 枚举硬编码修复
- 新增 signals.py 注册 post_delete 信号，防止文件残留
- case_material_sync 枚举值硬编码改为 MaterialCategory.CASE_MATERIAL

P1: 大文件拆分 + 架构违规修复
- display_mixin.py(912行) 拆出 archive_mixin.py(588行)
- folder_scan_service.py 拆出 file_hash_utils.py + quality_card_detector.py
- 移除 Service 层 @staticmethod
- ClientPaymentImageService 改为构造函数注入
- contract_oa_sync_service 跨模块顶层导入改为 lazy import

P2: 模型改进
- contract_payment_service Decimal/float 混用全部改为 Decimal
- ClientPaymentRecord 添加 updated_at 字段
- FinalizedMaterial.file_path 添加 db_index
- Invoice 添加 content_hash 去重字段

P3: 代码质量
- case_material_sync 批量排序改用 bulk_update
- ContractPayment.__str__ 显示合同名称
- ContractPayment.contract on_delete CASCADE 改为 PROTECT
