+++ backend/apps/contracts/apps.py
@@ -8,3 +8,6 @@ class ContractsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "apps.contracts"
     verbose_name = _("合同管理")
+
+    def ready(self) -> None:
+        import apps.contracts.signals
